requirements/requirements_dev.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==4.0.1
certifi==2024.12.14
click==8.1.8
coverage==7.6.10
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.115.7
freezegun==1.5.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
Mako==1.3.8
MarkupSafe==3.0.2
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
pyasn1==0.6.1
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.3
tomli==2.2.1
typing_extensions==4.12.2
uvicorn==0.34.0
